Automated pedestal calib, improved file access while running
- Removed 'Pedestal Calibration' button. Instead, this is done
automatically at the start of each event. This now means the dialogue
prompt suggesting to disconnect the signals doesn't happen. So, if the
probability of seeing an event in the digitizer acquisition is
significant, then this may lead to an error.
- The output files are now flushed after every run is completed, and are
closed upon opening a new file to avoid errors that can occur if the
files are copied before finishing a session.