Added on-the-fly data compression using HDF5 library, for digitizer data only
- This affects the attainable count rate, but compresses the output file
size by a factor of 2
- There's some code to handle the QDC data, but to use this I'd need to
handle complex dataspaces, and since the data size is small it isn't
worth the effort.
- The on-the fly compression is not the best solution. Compressing after
the fact to a .zip archive gives similar file sizes regardless of
whether on-the-fly compression was used. I should just modify my code to
put the data in a .zip archive following a run.